+++ packages/react-native-eyecandy/src/components/Avatar/Avatar.tsx
@@ -3,6 +3,7 @@ import {
   ImageBackground,
   ImageBackgroundProps,
   StyleSheet,
+  View,
 } from 'react-native';
 
 import { useTheme } from '@nomada-sh/react-native-eyecandy-theme';
@@ -11,23 +12,38 @@ import { ms } from '../../utils';
 
 export interface AvatarProps extends ImageBackgroundProps {
   size?: number;
+  fallbackComponent?: React.ReactNode;
+  fallback?: boolean;
 }
 
-function Avatar({ size = ms(64), style, ...props }: AvatarProps) {
+function Avatar({
+  size = ms(64),
+  style,
+  fallbackComponent,
+  fallback,
+  ...props
+}: AvatarProps) {
   const { dark, palette } = useTheme();
+
+  const dynamicStyles = StyleSheet.create({
+    container: {
+      width: size,
+      height: size,
+      borderRadius: size / 2,
+      backgroundColor: palette.grey[dark ? 800 : 100],
+    },
+  });
+
+  if (fallback && fallbackComponent)
+    return (
+      <View style={[dynamicStyles.container, styles.container, style]}>
+        {fallbackComponent}
+      </View>
+    );
 
   return (
     <ImageBackground
-      style={[
-        {
-          width: size,
-          height: size,
-          borderRadius: size / 2,
-          backgroundColor: palette.grey[dark ? 800 : 100],
-        },
-        styles.container,
-        style,
-      ]}
+      style={[dynamicStyles.container, styles.container, style]}
       {...props}
     />
   );