+++ packages/react-native-eyecandy/src/components/AvatarEdit/AvatarEdit.tsx
@@ -1,15 +1,7 @@
 import React from 'react';
-import {
-  GestureResponderEvent,
-  Pressable,
-  PressableProps,
-  StyleProp,
-  StyleSheet,
-  ViewStyle,
-  ImageSourcePropType,
-} from 'react-native';
+import { Pressable, StyleProp, StyleSheet, ViewStyle } from 'react-native';
 
-import { Camera, Photo } from '@nomada-sh/react-native-eyecandy-icons';
+import { Camera, Photo, Trash } from '@nomada-sh/react-native-eyecandy-icons';
 import { useColors } from '@nomada-sh/react-native-eyecandy-theme';
 import ReactNativeHapticFeedback from 'react-native-haptic-feedback';
 import {
@@ -17,37 +9,58 @@ import {
   ImagePickerResponse,
   launchCamera,
   launchImageLibrary,
+  Asset,
 } from 'react-native-image-picker';
 
 import ActionSheet, { ActionSheetOption } from '../ActionSheet';
 import Avatar, { AvatarProps } from '../Avatar';
 import IconButton from '../IconButton';
 
-interface OnChangeError {
+export interface AvatarEditError {
   code: string;
   message: string;
 }
 
-export interface AvatarEditProps extends Omit<PressableProps, 'style'> {
+export type AvatarEditAsset = Asset;
+
+export interface AvatarEditProps {
   size?: number;
   source: AvatarProps['source'];
+  avatarStyle?: StyleProp<ViewStyle>;
+  imageStyle?: AvatarProps['imageStyle'];
   style?: StyleProp<ViewStyle>;
-  onChange?: (image?: ImageSourcePropType, error?: OnChangeError) => void;
+  onChange?: (image?: AvatarEditAsset, error?: AvatarEditError) => void;
+  onDelete?: () => void;
+  editable?: boolean;
+  onPress?: () => void;
   fromGalleryText?: string;
   fromCameraText?: string;
+  deleteText?: string;
+  cancelText?: string;
   title?: string;
+  fallbackComponent?: React.ReactNode;
+  fallback?: boolean;
 }
 
-function AvatarEdit({
+// TODO: Improve props.
+
+export function AvatarEdit({
   style,
   size = 100,
   source,
   onPress: onPressProp,
   onChange,
+  onDelete,
   fromGalleryText = 'From Gallery',
   fromCameraText = 'From Camera',
+  deleteText = 'Delete Avatar',
   title = 'Change Avatar',
-  ...props
+  cancelText = 'Cancel',
+  editable = true,
+  imageStyle,
+  avatarStyle,
+  fallback,
+  fallbackComponent,
 }: AvatarEditProps) {
   const colors = useColors(c => c.background.default);
   const [visible, setVisible] = React.useState(false);
@@ -71,8 +84,14 @@ function AvatarEdit({
       mediaType: 'photo',
     };
 
+    if (index === 2) {
+      onDelete?.();
+      return;
+    }
+
     const launch = index === 0 ? launchImageLibrary : launchCamera;
 
+    // TODO: Improve error handling.
     launch(options).then(onImageSelected).catch(console.error);
   };
 
@@ -85,16 +104,21 @@ function AvatarEdit({
       label: fromCameraText,
       icon: Camera,
     },
+    {
+      label: deleteText,
+      icon: Trash,
+    },
   ];
 
-  const onPress = (e: GestureResponderEvent) => {
+  const onPress = () => {
     ReactNativeHapticFeedback.trigger('impactLight');
     setVisible(true);
-    onPressProp && onPressProp(e);
+    onPressProp && onPressProp();
   };
 
   return (
     <Pressable
+      disabled={!editable}
       onPress={onPress}
       style={[
         {
@@ -104,29 +128,40 @@ function AvatarEdit({
         },
         style,
       ]}
-      {...props}
     >
-      <Avatar size={size} source={source} />
-      <IconButton
-        style={[
-          {
-            borderWidth: 1.5,
-            borderColor: colors.content,
-          },
-          styles.cameraIconButton,
-        ]}
-        icon={Camera}
-        iconSize={size * 0.14}
-        size={size * 0.24}
-        color="primary"
-        onPress={onPress}
+      <Avatar
+        style={avatarStyle}
+        size={size}
+        source={source}
+        imageStyle={imageStyle}
+        fallback={fallback}
+        fallbackComponent={fallbackComponent}
       />
+      {editable ? (
+        <IconButton
+          style={[
+            {
+              borderWidth: 1.5,
+              borderColor: colors.container,
+            },
+            styles.cameraIconButton,
+          ]}
+          icon={Camera}
+          iconSize={size * 0.2}
+          size={size * 0.3}
+          color="primary"
+          onPress={onPress}
+          disabled
+          hideDisabledOverlay
+        />
+      ) : null}
       <ActionSheet
         visible={visible}
         onClose={() => setVisible(false)}
         onPressAction={onPressAction}
         options={options}
         title={title}
+        cancelText={cancelText}
         showCancelIcon
       />
     </Pressable>
@@ -140,5 +175,3 @@ const styles = StyleSheet.create({
     right: 2,
   },
 });
-
-export default AvatarEdit;