+++ packages/react-native-eyecandy/src/components/AvatarEdit/index.ts
@@ -1,4 +1,1 @@
-import AvatarEdit from './AvatarEdit';
-
 export * from './AvatarEdit';
-export default AvatarEdit;